particles: NRPIC Stage 3d(b) -- style: cpplint+flake8 gates green, no functional change
Brings `python3 run_test_suite.py --style` to green repo-wide: 405 findings -> 0.

- 231 cpplint findings in the NRPIC particle files: 205 line reflows (112 in
  gr_boris.cpp); runtime/int -> <cstdint> uint64_t/int64_t for the conservation
  ledger and summary types, with the Allreduce moved to MPI_UINT64_T; <cstdio>
  include_what_you_use; read_particle.cpp include order; part_crossing.cpp braces.
- 15 more in NRPIC-touched outputs/pgen files (vtk_prtcl, restart, pgen.cpp,
  uni_mag_part): reflows.
- 46 upstream-inherited (zero NRPIC commits on those files): tabs + reflows +
  brace + includes in z4c/compact_object_tracker.cpp and z4c/z4c_calcrhs.cpp,
  reflows in pgen/z4c_spectre_mbh.cpp and z4c/z4c_Sbc.cpp, and main.cpp's
  <fenv.h> reordered + NOLINT (cpplint bans fenv.h outright; needed for FPE traps).
- 3 pre-existing `}}` custom-check violations in gr_boris.cpp (case/switch
  closers split).
- 110 flake8 findings (gate's python half, setup.cfg max-line 90): 109 in
  scripts/particles/ (2-space -> 4-space reindent, line wraps, one E741 rename
  l -> ell in gen_geodesic_orbit.py ks_metric), 1 in vis/python/make_athdf.py.

Zero behavior change, machine-verified: C++ edits token-stream-identical to the
parent commit (comment-stripped, string-concatenation-aware); python edits
ast.dump-identical (modulo the documented E741 rename). Verified no-perturbation
on the rebuilt trees: particles pytest tier green (both modules), serial suite
96/96, MPI suite QUICK 154/154, destruction/GR suite QUICK 82/82 -- all bitwise
contracts stand. Note: z4c_spectre_mbh.cpp is not compilable on this machine on
any commit (requires spectre::Exporter); its reflow rests on the token proof.

+++ src/bvals/bvals_part.cpp
@@ -6,6 +6,8 @@
 //! \file bvals_part.cpp
 //! \brief
 
+#include <cstdint>
+#include <cstdio>
 #include <cstdlib>
 #include <iostream>
 #include <utility>
@@ -82,7 +84,7 @@ TaskStatus ParticlesBoundaryValues::SetNewPrtclGID() {
   // counts (exit/sphere/lapse)} and the destroyed-side {sum tag, sum tag^2} checksum
   // accumulators of the two-sided conservation ledger (written only when debug >= 1)
   DvceArray1D<int> dstc("pdest_cnt",4);                    // zero-initialized
-  DvceArray1D<unsigned long long> dsums("pdest_sums",2);   // zero-initialized
+  DvceArray1D<uint64_t> dsums("pdest_sums",2);             // zero-initialized
   auto &pdestl = destroylist;
   auto &drr = destroy_rec_r;
   auto &dri = destroy_rec_i;
@@ -179,7 +181,7 @@ TaskStatus ParticlesBoundaryValues::SetNewPrtclGID() {
       if (dbg > 0) {
         // destroyed-side checksums of the two-sided conservation ledger (cast BEFORE
         // multiplying: int tag*tag overflows at tag >= 46341)
-        unsigned long long t = static_cast<unsigned long long>(pi(PTAG,p));
+        uint64_t t = static_cast<uint64_t>(pi(PTAG,p));
         Kokkos::atomic_add(&dsums(0), t);
         Kokkos::atomic_add(&dsums(1), t*t);
       }
